+++ src/components/layout/index.tsx
@@ -4,7 +4,8 @@ import TopMarquee from './TopMarquee';
 import Footer from './Footer';
 import Flakes from './Flakes';
 import Menu from './Menu';
-import LogoCart from './LogoCart';
+import Logo from './Logo';
+import Cart from './Cart';
 
 const Layout = (props: PropsWithChildren) => {
   const { children } = props;
@@ -13,9 +14,10 @@ const Layout = (props: PropsWithChildren) => {
       <Flakes />
       <TopMarquee />
       <$.ContentsContainer>
-      <LogoCart />
-      <Menu/>
-      {children}
+        <Logo />
+        <Menu />
+        <Cart />
+        {children}
       </$.ContentsContainer>
       <Footer />
     </$.Wrap>
